Define type for cdata blobs in to_binary_nif
Currently, when printing to binary, exml has an enforced way to
implement escaping, that is, by escaping character by character, which,
in the case of many characters that need escaping, increases the binary
size dramatically, and also makes the binary less readable to the
debugging eye. So we need a mechanism to tell exml to use cdata when
desired.