Fix checking for maximum parsed element size
The original option named 'max_child_size' had two major issues:
1. It only checked incomplete (chunked) elements
2. It also checked the stream opening tag

Ad 1: Added a check for complete elements as well.
The check prevents constructing the Erlang term.

Ad 2: Modified the option name to 'max_element_size', and made it
check the size of the stream-opening tag as well as sizes of any
element. This makes sense because:
- One could (maliciously or not) send a huge stream-opening tag
- When infinite_stream is used, the elements are not children anyway,
  so the name was misleading.

Other changes:
- Updated error messages.
- The check for incomplete elements was failing too soon - one more
  character should be accepted. Now at most max_element-size-1
  characters are accepted, because we know that at least one more
  character will be parsed.