Fix elp error: make sure subelement/3 does not return cdata
This is tricky, but the result might actually be a cdata in the unlikely case
the cdata content matches some element name, like this:

  <elem>elem</elem>

Check it like this:

  > {ok, Elem} = exml:parse(<<"<elem>elem</elem>">>).
  ...
  > exml_query:subelement(Elem, <<"elem">>, default).
  #xmlcdata{content = <<"elem">>}

This is because #xmlel.name is just an index and both #xmlel{} and #xmlcdata{} records
do have a field with this index - it's just different fields.

So the spec should properly specify this function can technically return a cdata...
Or we can make sure it's a #xmlel{} by matching on Result.